refactor (global) : image 로직 분리 및 공통 코드 리팩토링
