+++ src/main/java/com/backend/immilog/global/exception/ErrorCode.java
@@ -4,5 +4,6 @@ import org.springframework.http.HttpStatus;
 
 public interface ErrorCode {
     HttpStatus getStatus();
+
     String getMessage();
 }
